+++ examples/lib/stories/input/joystick_advanced.dart
@@ -105,11 +105,11 @@ class JoystickAdvancedGame extends FlameGame with HasDraggables, HasTappables {
       style: TextStyle(color: BasicPalette.white.color),
     );
     speedText = TextComponent(
-      'Speed: 0',
+      text: 'Speed: 0',
       textRenderer: _regular,
     )..isHud = true;
     directionText = TextComponent(
-      'Direction: idle',
+      text: 'Direction: idle',
       textRenderer: _regular,
     )..isHud = true;