+++ packages/flame/test/text_test.dart
@@ -45,7 +45,7 @@ void main() {
     });
 
     test('change parameters of text component', () {
-      final tc = TextComponent<TextPaint>('foo');
+      final tc = TextComponent<TextPaint>(text: 'foo');
       tc.textRenderer = tc.textRenderer.copyWith(
         (c) => c.copyWith(fontSize: 200),
       );
@@ -55,12 +55,12 @@ void main() {
     test('custom renderer', () {
       TextRenderer.defaultRenderersRegistry[_CustomTextRenderer] =
           () => _CustomTextRenderer();
-      final tc = TextComponent<_CustomTextRenderer>('foo');
+      final tc = TextComponent<_CustomTextRenderer>(text: 'foo');
       expect(tc.textRenderer, isA<_CustomTextRenderer>());
     });
 
     test('text component size is set', () {
-      final t = TextComponent('foobar');
+      final t = TextComponent(text: 'foobar');
       expect(t.size, isNot(equals(Vector2.zero())));
     });
   });