+++ packages/flame/CHANGELOG.md
@@ -33,6 +33,7 @@
  - Removed methods `preRender()` and `postRender()` in `Component`
  - Use `FlameTester` everywhere where it makes sense in the tests
  - Improved `IsometricTileMap`
+ - Initialization of all `PositionComponent`s can be done from `onLoad` instead of the constructor
  - Rename `HasTappableComponents` to `HasTappables`
  - Rename `HasDraggableComponents` to `HasDraggables`
  - Rename `HasHoverableComponents` to `HasHoverableis`